address PR issue of ui and hash check
